Decompile the shift-right opcodes, and keep integers exact
An integer hash crashed on ushr. ishr and ushr were unimplemented in the
parser, the disassembler, the operand tables and the integer analysis -
ishl was there and its two siblings never had been. ShiftRightOperation
writes as >> for both, since HLSL takes the signed or unsigned behaviour
from the type of the value rather than from the operator.

That left three things wrong, all older than the shift opcodes:

    t0 >> 0.000000        the amount, 16, read as float bits
    1013904192            for 1013904223
    >> float2(8, 16)      a vector of amounts, as floats

The first was D3D10Opcode.IsInteger not listing the shifts, so the
immediate took the float path. The second is why ConstantNode carries the
integer alongside the float: a float cannot hold every 32 bit integer and
1013904223 is one of them. The third is the same typing one component
wider, which had been fixed once already and went out with the bathwater
when I reverted a whole file over the saturate_step experiment; re-applied
on its own this time.

Two fixtures improve on their own account: texture_load says
tx.Load(int3(1, 2, 0)), which is the signature, and clip_loop initialises
an integer register with int2.

+++ HlslDecompiler.Tests/ShaderSources/ps_4_0/texture_load.fx
@@ -2,5 +2,5 @@ Texture2D tx;
 
 float4 main() : SV_Target
 {
-	return tx.Load(float3(1, 2, 0));
+	return tx.Load(int3(1, 2, 0));
 }
